ingestion/pyproject.toml
[build-system]
requires = [
  "setuptools >= 63.0.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "ingestion"
version = "0.0.1"
dependencies = [
    "requests",
    'importlib-metadata; python_version<"3.10"',
    'pandas==2.2.2',
    'pyarrow==15.0.2',
    'numpy==1.26.4',
    'pyspark==3.5.0',
    'pyyaml==6.0.1',
    'delta-spark==3.0.0'
]

[project.scripts]
review-ingestion = "ingestion.cli:run"

[project.optional-dependencies]
dev = [
    "pytest==7.3.0"
]
